requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coverage==5.3
defusedxml==0.6.0
dj-rest-auth==1.1.2
Django==3.1.1
django-allauth==0.42.0
django-filter==2.4.0
django-simple-history==2.11.0
django-templated-mail==1.1.1
djangorestframework==3.12.1
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
